repopick: Fetch/Pull strictly from gerrit
Initially, repopick would attempt to find, fetch/pull, then cherry-pick (if applicable) from github.
If it failed, then it would move on to gerrit next.
We do not need it to target github but only gerrit so remove the uneceassary code.
This will eliminiate that unwanted github error that we all cannot stand.

Change-Id: Ib7be3c2fe6d6f604160d8702569d2f710f60ce67
Signed-off-by: SHM <stephenminugh@gmail.com>